requirements.txt
fastapi
uvicorn
praw
cloudinary
asyncpg
supabase
python-dotenv
httpx
python-multipart
pydantic 
email-validator
passlib
bcrypt
google-generativeai
python-jose[cryptography]